requirements.txt
streamlit
pandas
matplotlib
seaborn
statsmodels
numpy
psycopg2-binary
